Move unexported and shared functions to helpers.go
